+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-17T11:59:16.603Z",
+    "2026-06-17T13:32:19.477Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {